multiple layers of robot: logic, ui, render from logic
