+++ client/src/index.tsx
@@ -4,8 +4,7 @@ import injectSheet, { WithStylesProps } from 'react-jss';
 import classNames from 'classnames';
 import SquareContainer from './square-container';
 import Grid from './grid';
-import Robot from './robot';
-import { createDefaultRobot } from './robot/utils';
+import RobotUI from './robot/ui';
 import ControlPanel from './control-panel';
 import { useWorld, buildRobot } from './hooks';
 
@@ -46,6 +45,7 @@ const App = injectSheet(styles)(({
                     moveToLoc={moveToLoc}
                 />
             </div>
+            <RobotUI />
         </div>
     );
 });
